+++ frontend/_application/js/terrific-bootstrap.js
@@ -1,23 +1,20 @@
-
 // WP7 can't be detected with conditional comments.
 (navigator.userAgent.indexOf('Windows Phone') != -1 || navigator.userAgent.indexOf('WP7') != -1) && (document.documentElement.className += ' wp7');
+
 $(document.documentElement).removeClass('no-js');
 
 /**
  * Terrific Bootstrapping
  */
-(function (window, Tc, document) {
-	$(document).ready(function () {
-		var app
+(function (window, Tc) {
+	// If running inside a test frame, bail out.
+	if (window.xtc && window.xtc.isTest) { return; }
+	
+	$(function() {
+		var  app
 			,$body = $('body')
-			,config = {
-				dependencyPath: {
-					library: window.assetsUrl + '/scripts/libs-dyn/',
-					plugin:  window.assetsUrl + '/scripts/plugins-dyn/',
-					util:    window.assetsUrl + '/scripts/utils-dyn/'
-				}
-			},
-			moduleTest
+			,config = {}
+			,announceTcReady
 		;
 
 		// custom function for context selection
@@ -28,9 +25,15 @@ $(document.documentElement).removeClass('no-js');
 		app = new Tc.Application($body, config);
 		app.registerModules();
 		app.registerModule($body, 'PageController');
-		// after all modules are registered, register module tests
-		//typeof ModuleTest != 'undefined' && (moduleTest = new ModuleTest(app)); // Prepare atomic module tests
+
+		// After all modules are registered, register module tests
+		window.xtc && window.xtc.ModuleTests && (announceTcReady = new xtc.ModuleTests(app));
+
 		app.start();
-		//moduleTest && moduleTest.run();
+
+		// Tell xtc.ModuleTests that we're ready and hand it a callback, which it will call when it's ready too.
+		announceTcReady && announceTcReady(function(runTests) {
+			runTests();
+		});
 	});
-})(window, Tc, document);
+})(window, Tc);