+++ frontend/modules/example/test/example.test.js
@@ -1,23 +1,21 @@
-Tc.tests.Example = function( modName, $node, testApp ) {
+xtc.tests.Example = function(mod, options) {
 
 	 // Start a QUnit module for this Terrific module
-	module(modName, {
+	module(this.name, {
 		 // prepare something for each following tests
-		setup: function () {
-			 // Create a new instance of the Terrific module for each test() below.
-			var $ctx = $node.clone(true, true);
-			this.mod = testApp.registerModule($ctx, modName);
-			this.mod.start();
-		},
+		setup: function () {},
 		 // clean up after each test()
-		teardown: function () {
-			testApp.unregisterModules(this.mod);
-		}
+		teardown: function () {}
 	});
 
-	test( 'dummy tests', function() {
-
-		strictEqual( this.mod.$ctx.is('.skin-alternate'), false, 'Isn\'t CSS skin-alternate.' );
-		strictEqual( this.mod.$$('.debug').length, 1, 'Appends debug text.' );
+	test( 'Dummy', function() {
+		
+		if (!mod.$ctx.is('.skin-example-alternate')) {
+			strictEqual( mod.$$('.debug').length, 1, 'Appends debug text.' );
+		}
+		else {
+			ok( mod.$ctx.is('.skin-example-alternate'), 'Is .skin-example-alternate.' );
+			strictEqual( mod.$$('.debug').length, 2, 'Appends debug texts.' );
+		}
 	});
 };