+++ test/assets/fixtures/_application/js/application1.js
@@ -1,0 +1,3 @@
+(function application1() {
+	console.log(arguments.callee.name);
+})();