+++ test/assets/fixtures/_base/js/base1.js
@@ -1,0 +1,3 @@
+(function base1() {
+	console.log(arguments.callee.name);
+})();