+++ test/assets/fixtures/_inline/js/inline1.js
@@ -1,0 +1,3 @@
+(function inline1() {
+	console.log(arguments.callee.name);
+})();