+++ test/assets/fixtures/_inline/js/lib/inline-lib1.js
@@ -1,0 +1,3 @@
+(function inlineLib1() {
+	console.log(arguments.callee.name);
+})();