+++ test/assets/fixtures/modules/module1/module1.js
@@ -1,0 +1,46 @@
+(function($) {
+	/**
+	 * Module1 module implementation.
+	 *
+	 * @author mdiethelm
+	 * @namespace Tc.Module
+	 * @class Default
+	 * @extends Tc.Module
+	 */
+	Tc.Module.Module1 = Tc.Module.extend({
+
+		/**
+		 * Initializes the Module1 module.
+		 *
+		 * @method init
+		 * @constructor
+		 * @param {jQuery|Zepto} $ctx the jquery context
+		 * @param {Sandbox} sandbox the sandbox to get the resources from
+		 * @param {String} modId the unique module id
+		 */
+		init: function($ctx, sandbox, modId) {
+			// call base constructor
+			this._super($ctx, sandbox, modId);
+		},
+
+		/**
+		 * Hook function to do all of your module stuff.
+		 *
+		 * @method on
+		 * @param {Function} callback function
+		 * @return void
+		 */
+		on: function(callback) {
+			callback();
+		},
+
+		/**
+		 * Hook function to trigger your events.
+		 *
+		 * @method after
+		 * @return void
+		 */
+		after: function() {
+		}
+	});
+})(Tc.$);