+++ test/fixtures/terrific-modules/config-annotate.js
@@ -1,6 +1,0 @@
-module.exports = {
-	annotateModules: {
-		development: true
-	}
-	,repository: 'https://github.com/'
-};