+++ CHANGELOG.md
@@ -10,6 +10,10 @@ xxxx-xx-xx
  - Replace Underscore.js with Lo-Dash. Eliminates need for utils.deepExtend.
  - Set any attribute on the Terrific module wrapper.
  - Ready to deploy to Heroku using Strongloop.
+ - Change skins directory name from 'skin' to 'skins'. Can be changed in config.
+ - Module tests for the current page run in the browser.
+ - Less files in `reference` folders are included with @import (reference): Only imports what is actually used.
+
 
 # 0.6.3